Refac : Patterns Components styled 분리
