+++ src/components/patterns/CardSmall/CardSmall.jsx
@@ -1,9 +1,13 @@
 import React from "react";
 import PropTypes from "prop-types";
-import styled from "styled-components";
 
-import { ImageBox } from "@/components";
-import { colors, fontSize, fontWeight, loadings } from "@/_shared";
+import {
+  Layout,
+  Thumbnail,
+  TextBox,
+  Title,
+  Highlight,
+} from "./CardSmall.styled";
 
 const THEME = {
   LIGHT: "light",
@@ -57,51 +61,3 @@ CardSmall.defaultProps = {
 };
 
 export default CardSmall;
-
-const titleColor = {
-  light: colors.gray900,
-  dark: colors.gray25,
-};
-
-const Layout = styled.div`
-  width: 220px;
-  user-select: none;
-
-  .thumbnail {
-    margin-bottom: 16px;
-  }
-`;
-
-const Thumbnail = styled(ImageBox)`
-  width: 100%;
-`;
-
-const TextBox = styled.div`
-  div {
-    display: inline-block;
-    border-radius: 4px;
-  }
-`;
-
-const Title = styled.div`
-  min-width: 160px;
-  min-height: ${fontSize.p};
-
-  font-weight: ${fontWeight.bold};
-  font-size: ${fontSize.sm};
-  color: ${(props) => titleColor[props.theme]};
-
-  ${(props) => props.isLoading && loadings[props.theme]};
-`;
-
-const Highlight = styled.div`
-  min-width: 70px;
-  min-height: ${fontSize.p};
-  margin: 8px 0px;
-
-  font-size: ${fontSize.xs};
-  font-weight: ${fontWeight.bold};
-  color: ${(props) => (props.isDisabled ? colors.gray400 : colors.blue100)};
-
-  ${(props) => props.isLoading && loadings[props.theme]};
-`;
