+++ src/components/patterns/Dropdown/DropdownSmall/DropdownSmall.jsx
@@ -1,9 +1,8 @@
 import React from "react";
 import PropTypes from "prop-types";
-import styled from "styled-components";
 
 import { DropdownItem } from "../DropdownItem";
-import { colors, shadows } from "@/_shared";
+import { Layout, MenuBox } from "./DropdownSmall.styled";
 
 const THEME = {
   DARK: "dark",
@@ -49,40 +48,3 @@ DropdownSmall.defaultProps = {
 };
 
 export default DropdownSmall;
-
-const borderColor = {
-  dark: colors.gray800,
-  light: colors.gray100,
-};
-
-const bgColor = {
-  dark: colors.gray900,
-  light: colors.white,
-};
-
-const boxShadow = {
-  dark: "2px 2px 4px rgba(0, 0, 0, 0.185)",
-  light: shadows.base,
-};
-
-const Layout = styled.div`
-  display: flex;
-  flex-direction: column;
-  align-items: stretch;
-  position: absolute;
-
-  width: 100px;
-  border-radius: 8px;
-  border: 1px solid ${(props) => borderColor[props.theme]};
-
-  background-color: ${(props) => bgColor[props.theme]};
-  box-shadow: ${(props) => boxShadow[props.theme]};
-`;
-
-const MenuBox = styled.div`
-  padding: 4px 0;
-
-  > div {
-    justify-content: center;
-  }
-`;
