+++ src/components/patterns/Editor/Editor.jsx
@@ -1,12 +1,11 @@
 import React from "react";
 import PropTypes from "prop-types";
-import styled from "styled-components";
 
 import ReactQuill, { Quill } from "react-quill";
 import MarkdownShortcuts from "quill-markdown-shortcuts";
 import "react-quill/dist/quill.snow.css";
 
-import { colors } from "@/_shared";
+import { Layout } from "./Editor.styled";
 
 const THEME = {
   DARK: "dark",
@@ -58,185 +57,3 @@ Editor.defaultProps = {
 };
 
 export default Editor;
-
-const bgColor = {
-  light: colors.white,
-  dark: colors.black,
-};
-
-const borderColor = {
-  light: colors.gray300,
-  dark: colors.gray700,
-};
-
-const toolsColor = {
-  light: colors.gray700,
-  dark: colors.gray200,
-};
-
-const textColor = {
-  light: colors.gray900,
-  dark: colors.gray25,
-};
-
-const Layout = styled.div`
-  .ql-container {
-    border: 1px solid ${(props) => borderColor[props.theme]};
-    border-radius: 0 0 8px 8px;
-    font-size: 1rem;
-  }
-
-  .ql-editor {
-    min-height: 200px;
-    padding: 1.25rem;
-    font-family: Pretendard;
-    color: ${(props) => textColor[props.theme]};
-    letter-spacing: -0.02rem;
-  }
-
-  .ql-toolbar.ql-snow {
-    padding: 1rem;
-    border: 1px solid ${(props) => borderColor[props.theme]};
-    border-radius: 8px 8px 0 0;
-    font-family: Pretendard;
-
-    .ql-stroke {
-      stroke: ${(props) => toolsColor[props.theme]};
-    }
-
-    .ql-fill {
-      fill: ${(props) => toolsColor[props.theme]};
-    }
-  }
-
-  .ql-snow .ql-picker {
-    color: ${(props) => toolsColor[props.theme]};
-  }
-
-  .ql-snow .ql-picker-options {
-    background-color: ${(props) => bgColor[props.theme]};
-  }
-
-  .ql-toolbar.ql-snow .ql-picker.ql-expanded .ql-picker-options {
-    border: 1px solid ${(props) => borderColor[props.theme]};
-    border-radius: 4px;
-  }
-
-  .ql-toolbar.ql-snow .ql-picker.ql-expanded .ql-picker-label {
-    border: 1px solid ${(props) => borderColor[props.theme]};
-  }
-
-  .ql-snow .ql-editor h1 {
-    font-size: 1.5rem;
-    font-weight: 700;
-    padding: 3px 0;
-  }
-  .ql-snow .ql-editor h2 {
-    font-size: 1.25rem;
-    font-weight: 700;
-    padding: 3px 0;
-  }
-  .ql-snow .ql-editor h3 {
-    font-size: 1.125em;
-    font-weight: 700;
-    line-height: 1.3;
-    padding: 3px 0;
-  }
-
-  .ql-snow .ql-editor p {
-    font-size: 1rem;
-    line-height: 1.5rem;
-    padding: 3px 0;
-  }
-
-  .ql-syntax {
-    font-family: "SFMono-Regular", Consolas, "Liberation Mono", Menlo, Courier,
-      monospace;
-  }
-
-  .ql-snow .ql-editor pre.ql-syntax {
-    background-color: #1c2023;
-  }
-
-  .ql-video {
-    width: 560px;
-    height: 315px;
-  }
-
-  .ql-snow.ql-toolbar button:hover,
-  .ql-snow .ql-toolbar button:hover,
-  .ql-snow.ql-toolbar button:focus,
-  .ql-snow .ql-toolbar button:focus,
-  .ql-snow.ql-toolbar button.ql-active,
-  .ql-snow .ql-toolbar button.ql-active,
-  .ql-snow.ql-toolbar .ql-picker-label:hover,
-  .ql-snow .ql-toolbar .ql-picker-label:hover,
-  .ql-snow.ql-toolbar .ql-picker-label.ql-active,
-  .ql-snow .ql-toolbar .ql-picker-label.ql-active,
-  .ql-snow.ql-toolbar .ql-picker-item:hover,
-  .ql-snow .ql-toolbar .ql-picker-item:hover,
-  .ql-snow.ql-toolbar .ql-picker-item.ql-selected,
-  .ql-snow .ql-toolbar .ql-picker-item.ql-selected {
-    color: ${colors.blue100} !important;
-  }
-  .ql-snow.ql-toolbar button:hover .ql-fill,
-  .ql-snow .ql-toolbar button:hover .ql-fill,
-  .ql-snow.ql-toolbar button:focus .ql-fill,
-  .ql-snow .ql-toolbar button:focus .ql-fill,
-  .ql-snow.ql-toolbar button.ql-active .ql-fill,
-  .ql-snow .ql-toolbar button.ql-active .ql-fill,
-  .ql-snow.ql-toolbar .ql-picker-label:hover .ql-fill,
-  .ql-snow .ql-toolbar .ql-picker-label:hover .ql-fill,
-  .ql-snow.ql-toolbar .ql-picker-label.ql-active .ql-fill,
-  .ql-snow .ql-toolbar .ql-picker-label.ql-active .ql-fill,
-  .ql-snow.ql-toolbar .ql-picker-item:hover .ql-fill,
-  .ql-snow .ql-toolbar .ql-picker-item:hover .ql-fill,
-  .ql-snow.ql-toolbar .ql-picker-item.ql-selected .ql-fill,
-  .ql-snow .ql-toolbar .ql-picker-item.ql-selected .ql-fill,
-  .ql-snow.ql-toolbar button:hover .ql-stroke.ql-fill,
-  .ql-snow .ql-toolbar button:hover .ql-stroke.ql-fill,
-  .ql-snow.ql-toolbar button:focus .ql-stroke.ql-fill,
-  .ql-snow .ql-toolbar button:focus .ql-stroke.ql-fill,
-  .ql-snow.ql-toolbar button.ql-active .ql-stroke.ql-fill,
-  .ql-snow .ql-toolbar button.ql-active .ql-stroke.ql-fill,
-  .ql-snow.ql-toolbar .ql-picker-label:hover .ql-stroke.ql-fill,
-  .ql-snow .ql-toolbar .ql-picker-label:hover .ql-stroke.ql-fill,
-  .ql-snow.ql-toolbar .ql-picker-label.ql-active .ql-stroke.ql-fill,
-  .ql-snow .ql-toolbar .ql-picker-label.ql-active .ql-stroke.ql-fill,
-  .ql-snow.ql-toolbar .ql-picker-item:hover .ql-stroke.ql-fill,
-  .ql-snow .ql-toolbar .ql-picker-item:hover .ql-stroke.ql-fill,
-  .ql-snow.ql-toolbar .ql-picker-item.ql-selected .ql-stroke.ql-fill,
-  .ql-snow .ql-toolbar .ql-picker-item.ql-selected .ql-stroke.ql-fill {
-    fill: ${colors.blue100} !important;
-  }
-  .ql-snow.ql-toolbar button:hover .ql-stroke,
-  .ql-snow .ql-toolbar button:hover .ql-stroke,
-  .ql-snow.ql-toolbar button:focus .ql-stroke,
-  .ql-snow .ql-toolbar button:focus .ql-stroke,
-  .ql-snow.ql-toolbar button.ql-active .ql-stroke,
-  .ql-snow .ql-toolbar button.ql-active .ql-stroke,
-  .ql-snow.ql-toolbar .ql-picker-label:hover .ql-stroke,
-  .ql-snow .ql-toolbar .ql-picker-label:hover .ql-stroke,
-  .ql-snow.ql-toolbar .ql-picker-label.ql-active .ql-stroke,
-  .ql-snow .ql-toolbar .ql-picker-label.ql-active .ql-stroke,
-  .ql-snow.ql-toolbar .ql-picker-item:hover .ql-stroke,
-  .ql-snow .ql-toolbar .ql-picker-item:hover .ql-stroke,
-  .ql-snow.ql-toolbar .ql-picker-item.ql-selected .ql-stroke,
-  .ql-snow .ql-toolbar .ql-picker-item.ql-selected .ql-stroke,
-  .ql-snow.ql-toolbar button:hover .ql-stroke-miter,
-  .ql-snow .ql-toolbar button:hover .ql-stroke-miter,
-  .ql-snow.ql-toolbar button:focus .ql-stroke-miter,
-  .ql-snow .ql-toolbar button:focus .ql-stroke-miter,
-  .ql-snow.ql-toolbar button.ql-active .ql-stroke-miter,
-  .ql-snow .ql-toolbar button.ql-active .ql-stroke-miter,
-  .ql-snow.ql-toolbar .ql-picker-label:hover .ql-stroke-miter,
-  .ql-snow .ql-toolbar .ql-picker-label:hover .ql-stroke-miter,
-  .ql-snow.ql-toolbar .ql-picker-label.ql-active .ql-stroke-miter,
-  .ql-snow .ql-toolbar .ql-picker-label.ql-active .ql-stroke-miter,
-  .ql-snow.ql-toolbar .ql-picker-item:hover .ql-stroke-miter,
-  .ql-snow .ql-toolbar .ql-picker-item:hover .ql-stroke-miter,
-  .ql-snow.ql-toolbar .ql-picker-item.ql-selected .ql-stroke-miter,
-  .ql-snow .ql-toolbar .ql-picker-item.ql-selected .ql-stroke-miter {
-    stroke: ${colors.blue100} !important;
-  }
-`;
