+++ src/components/patterns/Progress/Progress.styled.js
@@ -1,0 +1,6 @@
+import styled from "styled-components";
+
+export const Layout = styled.div`
+  display: flex;
+  justify-content: space-between;
+`;
