+++ src/components/patterns/Sort/Sort.jsx
@@ -1,10 +1,9 @@
 import React, { useState } from "react";
 import PropTypes from "prop-types";
-import styled from "styled-components";
 
+import { Layout, Button, Text } from "./Sort.styled";
 import { DropdownSmall } from "@/components";
 import { Icon } from "@/foundations";
-import { colors, fontSize, animations } from "@/_shared";
 
 const THEME = {
   LIGHT: "light",
@@ -69,33 +68,3 @@ Sort.defaultProps = {
 };
 
 export default Sort;
-
-const Layout = styled.div`
-  display: flex;
-  position: relative;
-
-  .dropdown {
-    top: 2rem;
-    left: -1.4rem;
-    z-index: 9999;
-    animation: ${animations.dropdown} 0.3s cubic-bezier(0.3, 0, 0, 1);
-  }
-`;
-
-const Button = styled.div`
-  display: flex;
-  align-items: center;
-  color: ${colors.gray500};
-
-  cursor: pointer;
-`;
-
-const Text = styled.div`
-  width: 100%;
-  margin-left: 8px;
-
-  font-size: ${fontSize.sm};
-  color: ${colors.gray500};
-
-  user-select: none;
-`;
